+++ libs/gameGraphics.cpp
@@ -269,7 +269,7 @@ void GameGraphics::RenderingThreadPool::draw_section(int start, int end)
         //float wallAngle = (std::atan(m_gameGraphics.m_halfWallHeight / distance) );
         //float screenWallHeight = g_screenHeight  * wallAngle / (m_verticalVisibleAngle * m_gameGraphics.m_halfWallHeight);
 
-        //--this version is faster, has easy texture mapping but locks th evertical view angle at 90° (45° up 45° down)--
+        //--this version is faster, has easy texture mapping but locks th evertical view angle at 90ï¿½ (45ï¿½ up 45ï¿½ down)--
         float screenWallHeight = (g_screenHeight / distance) * m_gameGraphics.m_halfWallHeight;
         float floorHeight = (g_screenHeight - screenWallHeight) / 2;
 
@@ -322,7 +322,7 @@ void GameGraphics::RenderingThreadPool::draw_section(int start, int end)
                 else
                     posOnWallSide = currRay.hitPos.y + m_gameGraphics.m_mapData.playerTransform.coords.y;
 
-                posOnWallSide -= std::floorf(posOnWallSide);
+                posOnWallSide -= std::floor(posOnWallSide);
 
                 texVStep = currentTexture->height() / screenWallHeight;
